+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 
 **Added**
 
+- #1549 Added registry profile for jQuery UI settings
 - #1544 Progress indicator for Batch listing
 - #1536 Integrated Setup and Profiles from senaite.lims
 - #1534 Integrate browser resources from senaite.lims
@@ -58,6 +59,7 @@ Changelog
 
 **Fixed**
 
+- #1550 Fix Uncaught TypeError in combogrid
 - #1542 Fix sporadical errors when contacts do not have a valid email address
 - #1540 Fix flushing CCEmail fields in Sample Add Form
 - #1533 Fix traceback from log when rendering stickers preview